fix lidar bug and add it to unitest
