+++ pgdrive/tests/test_object/vis_test_object.py
@@ -3,17 +3,18 @@ from pgdrive.envs.generation_envs.side_pass_env import SidePassEnv
 # setup_logger(True)
 
 
-def test_collision(render):
+def test_collision(render=False):
     env = SidePassEnv({"manual_control": render, "use_render": render, "debug": True})
     o = env.reset()
 
     for i in range(1, 100000 if render else 2000):
         o, r, d, info = env.step([0, 1])
-        env.render()
-        # if d:
-        #     break
+        if render:
+            env.render()
+        if d:
+            break
     env.close()
 
 
 if __name__ == "__main__":
-    test_collision(True)
+    test_collision(False)
